chore: rename and align agent name for coagents-qa
