+++ examples/coagents-qa/agent/email_agent/demo.py
@@ -8,7 +8,7 @@ from fastapi import FastAPI
 import uvicorn
 from copilotkit.integrations.fastapi import add_fastapi_endpoint
 from copilotkit import CopilotKitRemoteEndpoint, LangGraphAgent
-from my_agent.agent import graph
+from email_agent.agent import graph
 
 
 app = FastAPI()
@@ -28,7 +28,7 @@ def main():
     """Run the uvicorn server."""
     port = int(os.getenv("PORT", "8000"))
     uvicorn.run(
-        "my_agent.demo:app",
+        "email_agent.demo:app",
         host="0.0.0.0",
         port=port,
         reload=True,
